Move card content generation to separate file
